Companion for Substrate#9410, 9462
