+++ frame/staking/src/substrate_tests.rs
@@ -29,7 +29,7 @@
 //! explaining why you delete them.
 
 // --- paritytech ---
-use frame_election_provider_support::{ElectionProvider, Support};
+use frame_election_provider_support::*;
 use frame_support::{
 	assert_noop, assert_ok,
 	dispatch::WithPostDispatchInfo,
@@ -42,10 +42,7 @@ use sp_runtime::{
 	traits::{BadOrigin, Dispatchable, Zero},
 	Perbill, Percent,
 };
-use sp_staking::{
-	offence::{OffenceDetails, OnOffenceHandler},
-	SessionIndex,
-};
+use sp_staking::{offence::*, *};
 use substrate_test_utils::assert_eq_uvec;
 // --- darwinia-network ---
 use crate::{
@@ -311,7 +308,7 @@ fn rewards_should_work() {
 			);
 			assert_eq!(
 				*staking_events().last().unwrap(),
-				Event::EraPayout(0, total_payout_0, maximum_payout - total_payout_0)
+				Event::EraPaid(0, total_payout_0, maximum_payout - total_payout_0)
 			);
 			make_all_reward_payment(0);
 
@@ -348,7 +345,7 @@ fn rewards_should_work() {
 				maximum_payout * 2 - total_payout_0 - total_payout_1,
 				MICRO,
 			);
-			if let Event::EraPayout(a, b, c) = *staking_events().last().unwrap() {
+			if let Event::EraPaid(a, b, c) = *staking_events().last().unwrap() {
 				assert_eq!(a, 1);
 				assert_eq!(b, total_payout_1);
 				assert_eq_error_rate!(c, maximum_payout - total_payout_1, MICRO);
@@ -4675,7 +4672,7 @@ mod election_data_provider {
 					45
 				);
 				assert_eq!(staking_events().len(), 1);
-				assert_eq!(*staking_events().last().unwrap(), Event::StakingElection);
+				assert_eq!(*staking_events().last().unwrap(), Event::StakersElected);
 
 				for b in 21..45 {
 					run_to_block(b);
@@ -4692,7 +4689,7 @@ mod election_data_provider {
 					70
 				);
 				assert_eq!(staking_events().len(), 3);
-				assert_eq!(*staking_events().last().unwrap(), Event::StakingElection);
+				assert_eq!(*staking_events().last().unwrap(), Event::StakersElected);
 
 				Staking::force_no_eras(Origin::root()).unwrap();
 				assert_eq!(
@@ -4730,7 +4727,7 @@ mod election_data_provider {
 					55 + 25
 				);
 				assert_eq!(staking_events().len(), 6);
-				assert_eq!(*staking_events().last().unwrap(), Event::StakingElection);
+				assert_eq!(*staking_events().last().unwrap(), Event::StakersElected);
 				// The new era has been planned, forcing is changed from `ForceNew` to `NotForcing`.
 				assert_eq!(<ForceEra<Test>>::get(), Forcing::NotForcing);
 			})
